+++ src/TokenSchedule/ScheduleItem.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Numerics;
 using FluentValidation;
 using TokenSchedule.FluentValidation;
 using TokenSchedule.FluentValidation.Models;
@@ -7,11 +8,11 @@ namespace TokenSchedule
 {
     public class ScheduleItem : IValidatedScheduleItem
     {
-        public decimal Ratio { get; }
+        public BigInteger Ratio { get; }
         public DateTime StartDate { get; }
         public DateTime? FinishDate { get; }
 
-        public ScheduleItem(decimal ratio, DateTime startDate, DateTime? finishDate = null)
+        public ScheduleItem(BigInteger ratio, DateTime startDate, DateTime? finishDate = null)
         {
             Ratio = ratio;
             StartDate = startDate;
